refactor(frontend): move osmPoiMap into features/map with standard layout
Move shared/components/osmPoiMap/ into features/map/ with proper
feature structure (components/, composables/, types/, utils/).
Update all imports across browse, posts, and test files.

Also decouple useBrowseViewModel from browseFiltersStore — the
composable no longer reads selectedTagIds directly; the view
already owns that relationship. Move leaflet CSS import from
composable to OsmPoiMap.vue component where it belongs.
