+++ apps/frontend/src/features/browse/views/BrowseProfiles.vue
@@ -8,12 +8,12 @@ import { useOwnerProfileStore } from '@/features/myprofile/stores/ownerProfileSt
 import { useProfilesViewModel } from '../composables/useProfilesViewModel'
 import { useBrowseViewModel } from '../composables/useBrowseViewModel'
 import { useBrowseFiltersStore } from '@/features/browse/stores/browseFiltersStore'
-import { isValidLatLng, toLatLng } from '@/features/shared/components/osmPoiMap/mapUtils'
+import { isValidLatLng, toLatLng } from '@/features/map/utils/mapUtils'
 import { useDetailRouteState } from '@/features/shared/composables/useDetailRouteState'
 import { useDetailPanel } from '@/features/app/composables/useDetailPanel'
 
 import BrowseFilterBar from '../components/BrowseFilterBar.vue'
-import OsmPoiMap from '@/features/shared/components/osmPoiMap/OsmPoiMap.vue'
+import OsmPoiMap from '@/features/map/components/OsmPoiMap.vue'
 import NoResultsCTA from '../components/NoResultsCTA.vue'
 import ProfileMapCard from '../components/ProfileMapCard.vue'
 import PublicProfileView from '@/features/publicprofile/components/PublicProfileView.vue'
