+++ apps/frontend/src/features/map/types/map.types.ts
