+++ apps/frontend/src/features/map/utils/__tests__/mapUtils.spec.ts
